docs(2D:Interface): Ajoute documentation aux fonctions de gestion des markers
+ Refacto --> les fonction de gestion des overlays sont dans OlMapLayers
